Moved FMChange to trend package
