+++ src/main/java/com/mivek/parser/TAFParser.java
@@ -5,7 +5,6 @@ import java.time.LocalTime;
 import com.mivek.model.Airport;
 import com.mivek.model.BeginningValidity;
 import com.mivek.model.Cloud;
-import com.mivek.model.FMChange;
 import com.mivek.model.PROBChange;
 import com.mivek.model.TAF;
 import com.mivek.model.TEMPOChange;
@@ -15,6 +14,7 @@ import com.mivek.model.Visibility;
 import com.mivek.model.WeatherCondition;
 import com.mivek.model.trend.AbstractWeatherChange;
 import com.mivek.model.trend.BECMGChange;
+import com.mivek.model.trend.FMChange;
 import com.mivek.utils.Converter;
 import com.mivek.utils.Regex;
