+++ src/test/java/com/mivek/parser/TAFParserTest.java
@@ -23,13 +23,13 @@ import com.mivek.enums.Descriptive;
 import com.mivek.enums.Intensity;
 import com.mivek.enums.Phenomenon;
 import com.mivek.model.BeginningValidity;
-import com.mivek.model.FMChange;
 import com.mivek.model.TAF;
 import com.mivek.model.TEMPOChange;
 import com.mivek.model.TemperatureDated;
 import com.mivek.model.Validity;
 import com.mivek.model.trend.AbstractWeatherChange;
 import com.mivek.model.trend.BECMGChange;
+import com.mivek.model.trend.FMChange;
 import com.mivek.utils.Converter;
 
 import internationalization.Messages;
